# Socify-Companion/lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: SOCIFY API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/register:
    post:
      operationId: register
      tags: [auth]
      summary: Register a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginBody"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
...
  /analytics/overview:
    get:
      operationId: getAnalyticsOverview
      tags: [analytics]
      summary: Get analytics overview metrics
      parameters:
        - name: workspaceId
          in: query
          required: true
          schema:
            type: integer
        - name: period
          in: query
          schema:
            type: string
            enum: [7d, 30d, 90d]
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AnalyticsOverview"
  /analytics/engagement:
    get:
      operationId: getEngagementTimeline
      tags: [analytics]
      summary: Get engagement timeline data
      parameters:
        - name: workspaceId
          in: query
          required: true
          schema:
            type: integer
        - name: period
          in: query
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TimelinePoint"
  /analytics/top-posts:
    get:
      operationId: getTopPosts
      tags: [analytics]
      summary: Get top performing posts
      parameters:
        - name: workspaceId
          in: query
          required: true
          schema:
            type: integer
        - name: limit
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Post"
  /analytics/platform-breakdown:
    get:
      operationId: getPlatformBreakdown
      tags: [analytics]
      summary: Get per-platform analytics breakdown
      parameters:
        - name: workspaceId
          in: query
          required: true
          schema:
            type: integer
        - name: period
          in: query
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PlatformStat"

  # ─── LEADS / CRM ────────────────────────────────────────────────────────────
...
  /ai/generate-caption:
    post:
      operationId: generateCaption
      tags: [ai]
      summary: AI generate caption
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GenerateCaptionBody"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GenerateCaptionResponse"
  /ai/suggest-hashtags:
    post:
      operationId: suggestHashtags
      tags: [ai]
      summary: AI suggest hashtags
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SuggestHashtagsBody"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuggestHashtagsResponse"
components:
  schemas:
...
    RegisterBody:
      type: object
      properties:
        name:
          type: string
        email:
          type: string
        password:
          type: string
      required: [name, email, password]
    LoginBody:
      type: object
      properties:
        email:
          type: string
        password:
          type: string
      required: [email, password]
    AuthResponse:
      type: object
      properties:
        token:
          type: string
        user:
          $ref: "#/components/schemas/User"
      required: [token, user]
    User:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        avatar:
          type: string
          nullable: true
        role:
          type: string
          enum: [superadmin, user]
        currentWorkspaceId:
          type: integer
          nullable: true
        createdAt:
          type: string
      required: [id, name, email, role, createdAt]

    # ── Workspace ──
...
    Post:
      type: object
      properties:
        id:
          type: integer
        workspaceId:
          type: integer
        authorId:
          type: integer
        authorName:
          type: string
        caption:
          type: string
        mediaUrls:
          type: array
          items:
            type: string
        mediaType:
          type: string
          enum: [image, video, carousel, text, reel, story]
        platforms:
          type: array
          items:
            type: string
        hashtags:
          type: array
          items:
            type: string
        status:
          type: string
          enum: [draft, pending_approval, revision_requested, approved, scheduled, publishing, published, failed, archived]
        scheduledAt:
          type: string
          nullable: true
        publishedAt:
          type: string
          nullable: true
        totalReach:
          type: integer
          nullable: true
        totalEngagement:
          type: integer
          nullable: true
        campaignId:
          type: integer
          nullable: true
        aiGenerated:
          type: boolean
        createdAt:
          type: string
      required: [id, workspaceId, authorId, authorName, caption, mediaUrls, mediaType, platforms, hashtags, status, aiGenerated, createdAt]
...
    AnalyticsOverview:
      type: object
      properties:
        totalFollowers:
          type: integer
        followerGrowth:
          type: number
        totalReach:
          type: integer
        reachGrowth:
          type: number
        totalEngagement:
          type: integer
        engagementGrowth:
          type: number
        totalPosts:
          type: integer
        postsGrowth:
          type: number
        avgEngagementRate:
          type: number
        clickThrough:
          type: number
      required: [totalFollowers, followerGrowth, totalReach, reachGrowth, totalEngagement, engagementGrowth, totalPosts, postsGrowth, avgEngagementRate, clickThrough]
    TimelinePoint:
      type: object
      properties:
        date:
          type: string
        reach:
          type: integer
        engagement:
          type: integer
        followers:
          type: integer
        impressions:
          type: integer
      required: [date, reach, engagement, followers, impressions]
    PlatformStat:
      type: object
      properties:
        platform:
          type: string
        followers:
          type: integer
        reach:
          type: integer
        engagement:
          type: integer
        posts:
          type: integer
        engagementRate:
          type: number
      required: [platform, followers, reach, engagement, posts, engagementRate]

    # ── Lead ──
...
    GenerateCaptionBody:
      type: object
      properties:
        topic:
          type: string
        platform:
          type: string
        tone:
          type: string
        includeHashtags:
          type: boolean
      required: [topic, platform, tone]
    GenerateCaptionResponse:
      type: object
      properties:
        caption:
          type: string
        hashtags:
          type: array
          items:
            type: string
      required: [caption, hashtags]
    SuggestHashtagsBody:
      type: object
      properties:
        caption:
          type: string
        platform:
          type: string
      required: [caption, platform]
    SuggestHashtagsResponse:
      type: object
      properties:
        hashtags:
          type: array
          items:
            type: string
      required: [hashtags]